Consolidate imports: use store.tsx as single entry point
All files now import from './store' instead of './graph-store'.
store.tsx re-exports useGraphStore and helpers from graph-store.ts,
providing a single public API while keeping graph-store.ts as an
implementation detail.

Tests: 3131/3131 passing
