+++ noodles-editor/src/noodles/fields.ts
@@ -8,7 +8,7 @@ import { colorToHex } from '../utils/color'
 import { debugSetValue } from '../utils/debug'
 import type { BetterDeckProps, BetterMapProps } from '../visualizations'
 import type { inputComponents } from './components/field-components'
-import { useGraphStore } from './graph-store'
+import { useGraphStore } from './store'
 import type { IOperator, Operator } from './operators'
 import type { DeckViewDescriptor, DeckViewValue } from './types'
 import { deepEqual } from './utils/deep-equal'
